refactor: moved spinner to a separate folder
